Move SplitString transformation to newest string group
