+++ src/Refinery/String/Transformation/SplitString.php
@@ -1,7 +1,7 @@
 <?php
 /* Copyright (c) 2017 Stefan Hecken <stefan.hecken@concepts-and-training.de> Extended GPL, see docs/LICENSE */
 
-namespace ILIAS\Refinery\Transformation\Transformations;
+namespace ILIAS\Refinery\String\Transformation;
 
 use ILIAS\Data\Factory;
 use ILIAS\Data\Result;
@@ -17,7 +17,7 @@ class SplitString implements Transformation {
 	protected $delimiter;
 
 	/**
-	 * @var DataFactory
+	 * @var Factory
 	 */
 	private $factory;
